+++ Leomard/Views/PreferenceView.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import SwiftUI
+import MarkdownUI
 
 fileprivate struct FrequencyOption: Hashable, Equatable {
     let name: String
@@ -20,8 +21,8 @@ struct PreferencesView: View {
         .init(name: "General", icon: "gearshape", color: .blue),
         .init(name: "Content", icon: "text.alignleft", color: .cyan),
         .init(name: "Display", icon: "display", color: .teal),
-        .init(name: "Updates", icon: "square.and.arrow.down.on.square", color: .green)
-        //.init(name: "Experimental", icon: "testtube.2", color: .red)
+        .init(name: "Updates", icon: "square.and.arrow.down.on.square", color: .green),
+        .init(name: "Experiments", icon: "testtube.2", color: .red)
     ]
     @State fileprivate var currentSelection: PreferenceOption?
     
@@ -37,6 +38,9 @@ struct PreferencesView: View {
     @State var manuallyCheckedForUpdate: Bool = false
     
     @State fileprivate var selectedViewType: ViewModeOption = .singleColumn
+    
+    @State var helpText: String = ""
+    @State var showHelp: Bool = false
     
     @Environment(\.colorScheme) var colorScheme
     
@@ -55,7 +59,7 @@ struct PreferencesView: View {
         }
         .task {
             self.currentSelection = self.preferenceOptions[0]
-        
+            
             for notificationCheckFrequency in notificationCheckFrequencies {
                 if UserPreferences.getInstance.checkNotifsEverySeconds == notificationCheckFrequency.seconds {
                     selectedNotificaitonCheckFrequency = notificationCheckFrequency
@@ -67,6 +71,9 @@ struct PreferencesView: View {
             }
             
             self.selectedViewType = UserPreferences.getInstance.twoColumnView ? .twoColumns : .singleColumn
+        }
+        .overlay {
+            helpOverlay
         }
     }
     
@@ -130,8 +137,8 @@ struct PreferencesView: View {
                     displayPreferences
                 case preferenceOptions[3]:
                     updatesPreferences
-                //case preferenceOptions[4]:
-                    //experimentalPreferences
+                case preferenceOptions[4]:
+                    experimentalPreferences
                 default:
                     Text("")
                 }
@@ -151,9 +158,6 @@ struct PreferencesView: View {
             .onChange(of: selectedNotificaitonCheckFrequency) { value in
                 UserPreferences.getInstance.checkNotifsEverySeconds = value.seconds
             }
-            Text("Note: Notifications are not checked when app is closed.")
-                .frame(maxWidth: .infinity, alignment:.leading)
-                .lineLimit(nil)
         }
         VStack(alignment: .leading) {
             Text("Inbox")
@@ -166,7 +170,7 @@ struct PreferencesView: View {
                 Toggle("Posts", isOn: UserPreferences.getInstance.$preferDisplayNameCommunityPost)
                     .frame(maxWidth: .infinity, alignment: .leading)
                 Toggle("Followed", isOn: UserPreferences.getInstance.$preferDisplayNameCommunityFollowed)
-                        .frame(maxWidth: .infinity, alignment: .leading)
+                    .frame(maxWidth: .infinity, alignment: .leading)
             }
             Divider()
             VStack {
@@ -317,7 +321,65 @@ struct PreferencesView: View {
     @ViewBuilder
     private var experimentalPreferences: some View {
         VStack(alignment: .leading) {
-            // Unused for now.
+            HStack {
+                Toggle("Store liked posts locally", isOn: UserPreferences.getInstance.$saveLikedPosts)
+                Button("?", action: {
+                    self.helpText = """
+# Store liked posts locally
+Liked posts will be saved into your session, which then can be browsed in your Profile view under **Liked** posts.
+
+Leomard does not know which posts you liked in other apps, and will add them if either you see that post in, or you like one in Leomard only.
+
+The information about what you like is securely stored on your device.
+"""
+                    self.showHelp = true
+                })
+                .cornerRadius(360)
+            }
+        }
+    }
+    
+    @ViewBuilder
+    private var helpOverlay: some View {
+        if showHelp {
+            ZStack {
+                Color(white: 0, opacity: 0.33)
+                    .onTapGesture {
+                        showHelp = false
+                    }
+                    .ignoresSafeArea()
+                VStack {
+                    VStack(alignment: .leading) {
+                        HStack {
+                            Button("Dismiss", action: { showHelp = false })
+                                .buttonStyle(.link)
+                            Spacer()
+                                .buttonStyle(.link)
+                        }
+                        .frame(
+                            minWidth: 0,
+                            maxWidth: .infinity,
+                            alignment: .leading
+                        )
+                        .padding(.top, 10)
+                        .padding(.bottom, 0)
+                        let content = MarkdownContent(helpText)
+                        Markdown(content)
+                            .frame(maxHeight: .infinity)
+                    }
+                    .frame(maxWidth: .infinity, minHeight: 0, maxHeight: .infinity)
+                }
+                .padding()
+                .frame(maxWidth: 400, minHeight: 0, maxHeight: 250)
+                .background(Color(.textBackgroundColor))
+                .overlay(
+                    RoundedRectangle(cornerRadius: 8)
+                        .stroke(Color(.windowFrameTextColor), lineWidth: 1)
+                )
+                .cornerRadius(8)
+                .listStyle(SidebarListStyle())
+                .scrollContentBackground(.hidden)
+            }
         }
     }
 }